pyproject.toml
[project]
name = "example-amazon-affiliate-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "amightygirl-paapi5-python-sdk==1.0.0",
    "requests>=2.32.3",
]
